+++ R/filter_mnirs.R
@@ -6,104 +6,119 @@
 #'   2. A Butterworth digital filter.
 #'   3. A simple moving average.
 #'
-#' @param sample_rate A numeric value for the sample rate in Hz for
-#'   `method = "butterworth"`. Will be taken from metadata or estimated
-#'   from `time_channel` if not defined explicitly.
 #' @param method A character string indicating how to filter the data (see
 #'   *Details*).
 #'   \describe{
 #'      \item{`"smooth_spline"`}{Fits a cubic smoothing spline.}
-#'      \item{`"butterworth"`}{Uses a centred Butterworth digital filter.
-#'      `type` must be defined (see *Details*).}
+#'      \item{`"butterworth"`}{Uses a centred Butterworth digital filter.}
 #'      \item{`"moving_average"`}{Uses a centred moving average filter.}
 #'   }
-#' @param spar A numeric value defining the smoothing parameter for
-#'   `method = "smooth_spline"`.
-#' @param order An integer defining the filter order for
-#'   `method = "butterworth"` (*default* `order = 2`).
-#' @param W A one- or two-element numeric vector defining the filter cutoff
-#'   frequency(ies) for `method = "butterworth"`, as a fraction of the
-#'   Nyquist frequency (see *Details*).
-#' @param fc A one- or two-element numeric vector defining the filter cutoff
-#'   frequency(ies) for `method = "butterworth"`, in Hz (see *Details*).
-#' @param type A character string indicating the digital filter type for
-#'   `method = "butterworth"` (see *Details*).
-#'   \describe{
-#'      \item{`"low"`}{For a *low-pass* filter (the *default*).}
-#'      \item{`"high"`}{For a *high-pass* filter.}
-#'      \item{`"stop"`}{For a *stop-band* (band-reject) filter.}
-#'      \item{`"pass"`}{For a *pass-band* filter.}
-#'   }
-#' @param width An integer defining the local window in number of samples
-#'   around `idx` in which to perform the operation for
-#'   `method = "moving_average"`. Between
-#'   `[idx - floor(width/2), idx + floor(width/2)]`.
-#' @param span A numeric value defining the local window timespan around `idx`
-#'   in which to perform the operation for `method = "moving_average"`.
-#'   In units of `time_channel` or `t`, between `[t - span/2, t + span/2]`.
-#'
-#' @param na.rm A logical indicating whether missing values should be preserved
-#'   and passed through the filter (`TRUE`). Otherwise `FALSE` (the *default*)
-#'   will throw an error if there are any `NA`s (see *Details*).
-#' @param ... Additional arguments.
+#' @param na.rm A logical indicating whether missing values should be
+#'   ignored and passed through the filter (`TRUE`). Otherwise `FALSE`
+#'   (the *default*) may error if there are any `NA`s (see *Details*).
+#' @param ... Additional method-specific arguments must be specified 
+#'   (see *Details*).
 #' @inheritParams validate_mnirs
 #'
 #' @details
+#' ## filtering method
+#'
+#' #### `method = "smooth_spline"`
+#'
+#' Applies a non-parametric cubic smoothing spline from
+#' [stats::smooth.spline()]. Smoothing is defined by the parameter `spar`,
+#' which can be left as `NULL` and automatically determined via penalised
+#' log likelihood. This usually works well for responses occurring on the 
+#' order of minutes or longer. `spar` can be specified typically, but not 
+#' necessarily, in the range `spar = [0, 1]`.
+#'
+#' Additional arguments (`...`) accepted when `method = "smooth_spline"`:
+#'
 #' \describe{
-#'   \item{`method = "smooth_spline"`}{Applies a non-parametric cubic
-#'   smoothing spline from [stats::smooth.spline()]. Smoothing is defined
-#'   by the parameter `spar`, which can be left as `NULL` and automatically
-#'   determined via penalised log liklihood. This usually works well for
-#'   smoothing responses occurring on the order of minutes or longer. `spar`
-#'   can be defined explicitly, typically (but not necessarily) in the range
-#'   `spar = [0, 1]`.}
-#'
-#'   \item{`method = "butterworth"`}{Applies a centred (two-pass
-#'   symmetrical) Butterworth digital filter from [signal::butter()] and
-#'   [signal::filtfilt()].
-#'
-#'   Filter `type` defines how the desired signal frequencies are either
-#'   passed or rejected from the output signal. *Low-pass* and *high-pass*
-#'   filters allow only frequencies *lower* or *higher* than the cutoff
-#'   frequency `W` to be passed through as the output signal, respectively.
-#'   *Stop-band* defines a critical range of frequencies which are rejected
-#'   from the output signal. *Pass-band* defines a critical range of
-#'   frequencies which are passed through as the output signal.
-#'
-#'   The filter order (number of passes) is defined by `order`, typically in
-#'   the range `order = [1, 10]`. Higher filter order tends to capture more
-#'   rapid changes in amplitude, but also causes more distortion around
-#'   those change points in the signal. General advice is to use the
-#'   lowest filter order which sufficiently captures the desired rapid
-#'   responses in the data.
-#'
-#'   The critical (cutoff) frequency is defined by `W`, a numeric value for
-#'   *low-pass* and *high-pass* filters, or a two-element vector
-#'   `c(low, high)` defining the lower and upper bands for *stop-band* and
-#'   *pass-band* filters. `W` represents the desired fractional cutoff
-#'   frequency in the range `W = [0, 1]`, where `1` is the Nyquist
-#'   frequency, i.e., half the sample rate of the data in Hz.
-#'
-#'   Alternatively, the cutoff frequency can be defined by `fc` and
-#'   `sample_rate` together. `fc` represents the desired cutoff frequency
-#'   in Hz, and `sample_rate` is the sample rate of the recorded data in
-#'   Hz. `W = fc / (sample_rate / 2)`.
-#'
-#'   Only one of either `W` or `fc` should be defined. If both are defined,
-#'   `W` will be preferred over `fc`.}
-#'
-#'   \item{`method = "moving_average"`}{Applies a centred (symmetrical)
-#'   moving average filter in a local window, defined by either `width`
-#'   as the number of samples around `idx` between `[idx - floor(width/2),`
-#'   `idx + floor(width/2)]`. Or by `span` as the timespan in units of
-#'   `time_channel` between `[t - span/2, t + span/2]`. Specifying `width`
-#'   is often faster than `span`. A partial moving average will be calculated
-#'   at the edges of the data.}
+#'   \item{`spar`}{A numeric smoothing parameter passed to
+#'       [stats::smooth.spline()]. If `NULL` (*default*), automatically
+#'       determined via penalised log likelihood.}
 #' }
 #'
+#' #### `method = "butterworth"`
+#'
+#' Applies a centred (two-pass symmetrical) Butterworth digital filter
+#' from [signal::butter()] and [signal::filtfilt()].
+#'
+#' Filter `type` defines how the desired signal frequencies are either
+#' passed or rejected from the output signal. *Low-pass* and *high-pass*
+#' filters allow only frequencies *lower* or *higher* than the cutoff
+#' frequency, respectively to be passed through to the output signal.
+#' *Stop-band* defines a critical range of frequencies which are rejected
+#' from the output signal. *Pass-band* defines a critical range of
+#' frequencies which are passed through as the output signal.
+#'
+#' The filter order (number of passes) is defined by `order`, typically
+#' in the range `order = [1, 10]`. Higher filter order tends to capture
+#' more rapid changes in amplitude, but also causes more distortion
+#' around those change points in the signal. General advice is to use
+#' the lowest filter order which sufficiently captures the desired rapid
+#' responses in the data.
+#'
+#' The critical (cutoff) frequency can be defined by `W`, a numeric value
+#' for *low-pass* and *high-pass* filters, or a two-element vector
+#' `c(low, high)` defining the lower and upper bands for *stop-band*
+#' and *pass-band* filters. `W` represents the desired fractional cutoff
+#' frequency in the range `W = [0, 1]`, where `1` is the Nyquist
+#' frequency, i.e., half the `sample_rate` of the data in Hz.
+#'
+#' Alternatively, the cutoff frequency can be defined by `fc` and
+#' `sample_rate` together. `fc` represents the desired cutoff frequency
+#' directly in Hz, and `sample_rate` is the sample rate of the recorded data 
+#' in Hz. Where `W = fc / (sample_rate / 2)`.
+#'
+#' Only one of either `W` or `fc` should be defined. If both are
+#' defined, `W` will be preferred over `fc`.
+#'
+#' Additional arguments (`...`) accepted when `method = "butterworth"`:
+#'
+#' \describe{
+#'   \item{`order`}{An integer for the filter order (*default* `2`).}
+#'   \item{`W`}{A numeric fractional cutoff frequency within `[0, 1]`. One
+#'       of either `W` or `fc` must be specified.}
+#'   \item{`fc`}{A numeric absolute cutoff frequency in Hz. Used with
+#'       `sample_rate` to compute `W`.}
+#'   \item{`sample_rate`}{A numeric sample rate in Hz. Will be taken from
+#'       metadata or estimated from `time_channel` if not defined.}
+#'   \item{`type`}{A character string specifying filter type, one of:
+#'       `c("low", "high", "stop", "pass")` (`"low"` is the default).}
+#'   \item{`edges`}{A character string specifying the edge padding, one of:
+#'       `c("rev", "rep1", "none")` (`"rev"` is the default). 
+#'       See [filter_butter()].}
+#' }
+#'
+#' #### `method = "moving_average"`
+#'
+#' Applies a centred (symmetrical) moving average filter in a local
+#' window, defined by either `width` as the number of samples around
+#' `idx` between `[idx - floor(width/2), idx + floor(width/2)]`. Or by
+#' `span` as the timespan in units of `time_channel` between
+#' `[t - span/2, t + span/2]`.
+#'
+#' Additional arguments (`...`) accepted when `method = "moving_average"`:
+#'
+#' \describe{
+#'   \item{`width` or `span`}{Either an integer number of samples, or a
+#'       numeric time duration in units of `time_channel` within the local 
+#'       rolling window. One of either `width` or `span` must be specified.}
+#'   \item{`partial`}{Logical; `FALSE` by default, only returns values
+#'       where a full window of valid (non-`NA`) samples are available. 
+#'       If `TRUE`, ignores `NA` and allows calculation over partial windows 
+#'       at the edges of the data.}
+#' }
+#'
+#' ## missing values
+#'
 #' Missing values (`NA`) in `nirs_channels` will cause an error for
-#'   `method = "smooth_spline"` or `"butterworth"`, unless `na.rm = TRUE`.
-#'   Then `NA`s will be preserved and passed through in the returned data.
+#' `method = "smooth_spline"` or `"butterworth"`, unless `na.rm = TRUE`.
+#' Then `NA`s will be ignored and passed through to the returned data.
+#' For `method = "moving_average"`, `na.rm` controls whether `NA`s within
+#' each rolling window are removed before computing the mean.
 #'
 #' @returns
 #' A [tibble][tibble::tibble-package] of class *"mnirs"* with metadata
@@ -132,7 +147,7 @@
 #'     order = 2,              ## filter order number
 #'     W = 0.02,               ## filter fractional critical frequency `[0, 1]`
 #'     type = "low",           ## specify a "low-pass" filter
-#'     na.rm = TRUE            ## explicitly preserve NAs
+#'     na.rm = TRUE            ## explicitly ignore NAs
 #' )
 #' 
 #' ## note the smoothed `smo2` values
@@ -155,22 +170,14 @@ filter_mnirs <- function(
     data,
     nirs_channels = NULL,
     time_channel = NULL,
-    sample_rate = NULL,
     method = c("smooth_spline", "butterworth", "moving_average"),
-    spar = NULL,
-    order = 2,
-    W = NULL,
-    fc = NULL,
-    type = c("low", "high", "stop", "pass"),
-    width = NULL,
-    span = NULL,
     na.rm = FALSE,
     verbose = TRUE,
     ...
 ) {
     ## validation ====================================
     validate_mnirs_data(data)
-    ## using method aliases 
+    ## using method aliases
     method <- match.arg(method, c(
         "smooth_spline", "butterworth", "moving_average", "spline", "ma"
     ))
@@ -184,13 +191,10 @@ filter_mnirs <- function(
         verbose <- getOption("mnirs.verbose", default = TRUE)
     }
 
-    ## create object with class for method dispatch
-    data <- structure(
-        list(data = data),
-        class = c(method, "mnirs_filtered")
+    UseMethod(
+        "filter_mnirs", 
+        structure(data, class = c(method, "mnirs_filtered"))
     )
-
-    UseMethod("filter_mnirs", data)
 }
 
 
@@ -201,29 +205,20 @@ filter_mnirs.smooth_spline <- function(
     data,
     nirs_channels = NULL,
     time_channel = NULL,
-    sample_rate = NULL,
-    method = c("smooth_spline", "butterworth", "moving_average"),
-    spar = NULL,
-    order = 2,
-    W = NULL,
-    fc = NULL,
-    type = c("low", "high", "stop", "pass"),
-    width = NULL,
-    span = NULL,
+    method,
     na.rm = FALSE,
     verbose = TRUE,
     ...
 ) {
     ## validation ==========================================
-    metadata <- attributes(data)
-    ## verbose = FALSE because grouping irrelevant
     nirs_channels <- validate_nirs_channels(
         enquo(nirs_channels), data, verbose = FALSE
     )
     time_channel <- validate_time_channel(enquo(time_channel), data)
-    validate_numeric(
-        spar, 1, c(0, Inf), FALSE, msg1 = "one-element positive"
-    )
+    metadata <- attributes(data)
+    args <- list(...)
+    spar <- args$spar
+    validate_numeric(spar, 1, c(0, Inf), FALSE, msg1 = "one-element positive")
 
     ## processing ==========================================
     time_vec <- data[[time_channel]]
@@ -231,7 +226,7 @@ filter_mnirs.smooth_spline <- function(
     if (anyDuplicated(time_vec)) {
         cli_abort(c(
             "x" = "{.arg time_channel} has duplicated or irregular samples.",
-            "i" = "Re-sample with {.fn mnirs::resample_mnirs}."
+            "i" = "Re-sample first with {.fn mnirs::resample_mnirs}."
         ))
     }
 
@@ -247,7 +242,7 @@ filter_mnirs.smooth_spline <- function(
             cli_abort(c(
                 "x" = "{.arg nirs_channels} = {.val {.x}} contains internal \\
                 {.val {NA}}'s.",
-                "i" = "Set {.arg na.rm = TRUE} to preserve {.val {NA}}'s."
+                "i" = "Set {.arg na.rm = TRUE} to ignore {.val {NA}}'s."
             ))
         }
 
@@ -281,31 +276,28 @@ filter_mnirs.butterworth <- function(
     data,
     nirs_channels = NULL,
     time_channel = NULL,
-    sample_rate = NULL,
-    method = c("smooth_spline", "butterworth", "moving_average"),
-    spar = NULL,
-    order = 2,
-    W = NULL,
-    fc = NULL,
-    type = c("low", "high", "stop", "pass"),
-    width = NULL,
-    span = NULL,
+    method,
     na.rm = FALSE,
     verbose = TRUE,
     ...
 ) {
     ## validation ==========================================
-    metadata <- attributes(data)
-    ## verbose = FALSE because grouping irrelevant
     nirs_channels <- validate_nirs_channels(
         enquo(nirs_channels), data, verbose = FALSE
     )
     time_channel <- validate_time_channel(enquo(time_channel), data)
+    metadata <- attributes(data)
+    args <- list(...)
+    sample_rate <- args$sample_rate
     sample_rate <- validate_sample_rate(
         data, time_channel, sample_rate, verbose
     )
-    type <- match.arg(type)
-    edges <- list(...)$edges %||% "rev"
+    order <- args$order %||% 2L
+    W <- args$W
+    fc <- args$fc
+    type <- args$type %||% "low"
+    type <- match.arg(type, c("low", "high", "stop", "pass"))
+    edges <- args$edges %||% "rev"
 
     if (is.null(c(W, fc))) {
         cli_abort(c(
@@ -363,28 +355,21 @@ filter_mnirs.moving_average <- function(
     data,
     nirs_channels = NULL,
     time_channel = NULL,
-    sample_rate = NULL,
-    method = c("smooth_spline", "butterworth", "moving_average"),
-    spar = NULL,
-    order = 2,
-    W = NULL,
-    fc = NULL,
-    type = c("low", "high", "stop", "pass"),
-    width = NULL,
-    span = NULL,
+    method,
     na.rm = FALSE,
     verbose = TRUE,
     ...
 ) {
     ## validation ==========================================
-    metadata <- attributes(data)
-    ## verbose = FALSE because grouping irrelevant
     nirs_channels <- validate_nirs_channels(
         enquo(nirs_channels), data, verbose = FALSE
     )
     time_channel <- validate_time_channel(enquo(time_channel), data)
-    validate_width_span(width, span, verbose)
+    metadata <- attributes(data)
     args <- list(...)
+    width <- args$width
+    span <- args$span
+    partial <- args$partial %||% FALSE
 
     ## processing ==========================================
     time_vec <- data[[time_channel]]
@@ -395,9 +380,9 @@ filter_mnirs.moving_average <- function(
             t = time_vec,
             width = width,
             span = span,
+            partial = partial,
+            na.rm = na.rm,
             verbose = verbose,
-            args$partial %||% FALSE,
-            args$na.rm %||% FALSE,
             bypass_checks = TRUE
         )
     })
@@ -448,9 +433,9 @@ filter_mnirs.moving_average <- function(
 #'
 #' ## na.rm = FALSE (default): any NA in the window propagates to the result
 #' x_na <- c(1, NA, 3, 4, 5, NA, 7, 8)
-#' filter_ma(x_na, width = 3)
+#' filter_ma(x_na, width = 3, na.rm = FALSE)
 #'
-#' ## na.rm = TRUE: skip NAs within the window and return the mean of valid values
+#' ## na.rm = TRUE: skip NAs and return the local mean of local valid values
 #' filter_ma(x_na, width = 3, partial = TRUE, na.rm = TRUE)
 #'
 #' @rdname filter_ma
@@ -471,40 +456,52 @@ filter_ma <- function(
         if (missing(verbose)) {
             verbose <- getOption("mnirs.verbose", default = TRUE)
         }
-        validate_x_t(x, t)
-        validate_width_span(width, span, verbose)
     }
+    validate_x_t(x, t)
+    validate_width_span(width, span, verbose)
 
-    ## min_obs default to estimated width when span is specified
-    min_obs <- if (partial) {
-        1L
-    } else {
-        ## less strict span_width - 2 to allow start & end buffer
-        ## with irregular t values
-        max(width %||% (floor(span * estimate_sample_rate(t)) - 2L), 1L)
+    ## handle NAs
+    if (verbose && !na.rm && anyNA(x)) {
+        cli_warn(c(
+            "!" = "{.arg x} contains internal {.val {NA}}'s.",
+            "i" = "Set {.arg na.rm = TRUE} to ignore {.val {NA}}'s."
+        ))
     }
 
     ## processing ==============================================
     window_idx <- compute_local_windows(t, width = width, span = span)
 
-    ## check for min_obs
-    window_valid <- lapply(window_idx, \(.idx) which(is.finite(x)[.idx]))
-    window_idx[lengths(window_valid) < min_obs] <- NA_real_
-    if (verbose && all(is.na(window_idx))) {
-        cli_warn(c(
-            "!" = "Less than {.val {min_obs}} valid samples detected in \\
-            {.fn filter_ma} windows.",
-            "i" = "Specify {.arg width} >= {.val {2}} or increase \\
-            {.arg span} to include more samples."
-        ))
+    if (!partial) {
+        ## min_obs default to estimated width when span is specified
+        ## less strict span_width - 2 to allow start & end buffer
+        ## with irregular t values
+        min_obs <- max(
+            width %||% (floor(span * estimate_sample_rate(t)) - 2L),
+            1L
+        )
+
+        ## error if fewer valid samples than min_obs
+        if (sum(is.finite(x)) < min_obs) {
+            cli_abort(c(
+                "x" = "Insufficient samples detected.",
+                "i" = "Specify a smaller {.arg width} or {.arg span}, \\
+                or specify {.arg partial} = {.val {TRUE}}."
+            ))
+        }
+
+        which_partial <- lengths(window_idx) < min_obs
     }
 
     y <- vapply(window_idx, \(.idx) mean(x[.idx], na.rm = na.rm), numeric(1))
 
+    if (!partial) {
+        ## exclude incomplete windows (at edges)
+        y[which_partial] <- NA_real_
+    }
     ## NaN to NA
     y[!is.finite(y)] <- NA_real_
     return(y)
-}
+    }
 
 
 #' @rdname filter_ma
@@ -539,10 +536,21 @@ filter_moving_average <- function(
 #' of the data.
 #'
 #' @param x A numeric vector.
-#' @param edges A character string indicating how to pad `x` for edge detection.
+#' @param order An integer defining the filter order (*default* `order = 2`).
+#' @param W A one- or two-element numeric vector defining the filter cutoff
+#'   frequency(ies) as a fraction of the Nyquist frequency (see *Details*).
+#' @param type A character string indicating the digital filter type (see
+#'   *Details*).
 #'   \describe{
-#'      \item{`"rev"`}{(*the default*) Will pad `x` with the preceding 5% data
-#'      in reverse sequence.}
+#'      \item{`"low"`}{For a *low-pass* filter (the *default*).}
+#'      \item{`"high"`}{For a *high-pass* filter.}
+#'      \item{`"stop"`}{For a *stop-band* (band-reject) filter.}
+#'      \item{`"pass"`}{For a *pass-band* filter.}
+#'   }
+#' @param edges A character string indicating edge detection padding for `x`.
+#'   \describe{
+#'      \item{`"rev"`}{Will pad `x` with the preceding 5% data in reverse 
+#'      sequence (*the default*).}
 #'      \item{`"rep1"`}{Will pad `x` by repeating the last preceding value.}
 #'      \item{`"none"`}{Will return the unpadded [signal::filtfilt()] output.}
 #'   }
@@ -575,7 +583,7 @@ filter_moving_average <- function(
 #'   frequency, i.e., half the sample rate of the data in Hz.
 #'
 #' Missing values (`NA`) in `x` will cause an error unless `na.rm = TRUE`.
-#'   Then `NA`s will be preserved and passed through in the returned vector.
+#'   Then `NA`s will be ignored and passed through to the returned vector.
 #'
 #' @returns A numeric vector the same length as `x`.
 #'
@@ -614,11 +622,12 @@ filter_moving_average <- function(
 #' @export
 filter_butter <- function(
     x,
-    order = 2,
+    order = 2L,
     W,
     type = c("low", "high", "stop", "pass"),
     edges = c("rev", "rep1", "none"),
-    na.rm = FALSE
+    na.rm = FALSE,
+    ...
 ) {
     ## validation ============================================
     rlang::check_installed("signal", "to use Butterworth digital filter")
@@ -637,14 +646,14 @@ filter_butter <- function(
 
     ## processing ==============================================
     ## handle NAs
-    handle_na <- na.rm && any(is.na(x))
+    handle_na <- na.rm && anyNA(x)
     if (handle_na) {
         na_info <- preserve_na(x)
         x <- na_info$x_valid
-    } else if (any(is.na(x))) {
+    } else if (anyNA(x)) {
         cli_abort(c(
             "x" = "{.arg x} contains internal {.val {NA}}'s.",
-            "i" = "Set {.arg na.rm = TRUE} to preserve {.val {NA}}'s."
+            "i" = "Set {.arg na.rm = TRUE} to ignore {.val {NA}}'s."
         ))
     }
 
@@ -652,7 +661,7 @@ filter_butter <- function(
         y <- signal::filtfilt(signal::butter(n = order, W, type), x = x)
     } else {
         x_n <- length(x)
-        pad <- max(1, x_n %/% 20) ## 5% padded length
+        pad <- max(1L, x_n %/% 20) ## 5% padded length
 
         padded <- switch(
             edges,